WebStorage/tests/submissions/Infraware/storage_session_clear.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_session_clear</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>

		if (('sessionStorage' in window) && window.sessionStorage !== null){
			sessionStorage.clear();
			sessionStorage.setItem("name", "user1");

			if (sessionStorage.getItem("name") != null && sessionStorage.length == 1)
			{
				try {
					sessionStorage.clear();
				} catch (e) {
					test(function()
					{
						assert_unreached("clear method is failed.");
					}, "clear method test")
				}

				test(function() {
					assert_equals(sessionStorage.getItem("name"), null, "sessionStorage.getItem('name')")
				}, "result test of clear method (item)")

				test(function() {
					assert_equals(sessionStorage.length, 0, "sessionStorage.length")
				}, "result test of clear method (length)")

			} else {
				test(function() {
					assert_unreached("'name' item is not exist.");
				}, "item existence test")
			}
		} else {
			test(function()
			{
				assert_unreached("sessionStorage is not supported.");
			}, "sessionStorage support test")
		}	
	</script>

 </body>
</html>
